+++ packages/svelte-undp-components/src/lib/components/ui/Slider.svelte
@@ -57,7 +57,7 @@
 	}, 300);
 </script>
 
-<div class=" range-slider">
+<div class="range-slider">
 	<RangeSlider
 		bind:min
 		bind:max