+++ packages/svelte-undp-design/.storybook/main.ts
@@ -1,21 +1,16 @@
 import type { StorybookConfig } from '@storybook/sveltekit';
 const config: StorybookConfig = {
 	stories: ['../src/lib/**/*.mdx', '../src/lib/**/*.stories.@(js|jsx|ts|tsx|svelte)'],
-	addons: [
-		'@storybook/addon-links',
-		'@storybook/addon-essentials',
-		'@storybook/addon-interactions',
-		'@storybook/addon-svelte-csf',
-		'@storybook/addon-mdx-gfm'
-	],
+
+	addons: ['@storybook/addon-links', '@storybook/addon-svelte-csf', '@storybook/addon-docs'],
+
 	framework: {
 		name: '@storybook/sveltekit',
 		options: {}
 	},
-	docs: {
-		autodocs: true
-	},
+
 	staticDirs: ['../static'],
+
 	refs: {
 		'svelte-undp-components': {
 			title: 'Svelte UNDP Components',
@@ -26,6 +21,7 @@ const config: StorybookConfig = {
 			url: 'https://design.undp.org/'
 		}
 	},
+
 	core: {
 		disableTelemetry: true
 	}